api/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyJWT==2.1.0
tarantool==0.8.0